requirements.txt
albumentations
black
fastapi
joblib
matplotlib
numpy
pillow
pylint
python-multipart
requests
scikit-learn
scipy
ternausnet
torch
torchaudio
torchvision
tqdm
uvicorn
s3fs